requirements.txt
appdirs==1.4.0
Django==1.10.5
django-cors-headers==2.0.1
django-debug-toolbar==1.6
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
PyJWT==1.4.2
pyparsing==2.1.10
six==1.10.0
sqlparse==0.2.2
